pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["node_modules*", "*.js", "*.ts", "static*", "frontend*"]

[project]
name = "recording"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "instagrapi>=2.2.1",
    "pydantic-settings>=2.10.1",
    "pyotp>=2.9.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "typer>=0.19.2",
    "uvicorn[standard]>=0.35.0",
    "xai-sdk>=1.3.1",
]

[tool.pytest.ini_options]
pythonpath = "."

[project.scripts]
ignite = "app.main:main"

[tool.uv]
package = true
